Cleaning and splitting up the views.
I'm keeping the views as small as possible in order to make them easier
to change and new features easier to plug in.

I created a "bundles" dir containing a partial for each section of the
form and also removed head and javascripts from the main view since
these partials can became bigger and harder to maintain.