adding a phpmyadmin role for apache and nginx
Add the ability to use phpmyadmin on the dev box. To make it work you
only need to add a line in your /etc/hosts.
It's described in the usage documentation file.

The way it works now it would be easy to add a specific user and pass
for the phpmyadmin user.
If you think that it would be a good feature its just a matter of
adding the 2 field and putting their values in the var file.