Update windows.sh
Non existing /etc/ansible folder and missing sudo caused this error:

ERROR: The file ansible/inventories/dev is marked as executable, but failed to execute correctly. If this is not supposed to be an executable script, correct this with `chmod -x ansible/inventories/dev`.
Ansible failed to complete successfully. Any error output should be
visible above. Please fix these errors and try again.